+++ client/components/navbar.js
@@ -1,5 +1,4 @@
 import React from 'react'
-import {connect} from 'react-redux'
 import {Link} from 'react-router-dom'
 
 const Navbar = () => (
