+++ package.json
@@ -56,11 +56,8 @@
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-codemirror": "^1.0.0",
-
     "react-countdown": "^2.2.2",
-    "react-dom": "^16.8.6",
     "react-live-clock": "^5.0.6",
-
     "react-dom": "^16.13.1",
     "react-modal": "^3.11.2",
     "react-redux": "^7.0.1",
